+++ src/ReadonlyRecord.ts
@@ -101,9 +101,9 @@ export const collect = <K extends string, A, B>(f: (k: K, a: A) => B) => (
  * @category destructors
  * @since 3.0.0
  */
-export const toReadonlyArray: <K extends string, A>(
-  r: ReadonlyRecord<K, A>
-) => ReadonlyArray<readonly [K, A]> = collect((k, a) => [k, a])
+export const toReadonlyArray: <K extends string, A>(r: ReadonlyRecord<K, A>) => ReadonlyArray<readonly [K, A]> =
+  /*#__PURE__*/
+  collect((k, a) => [k, a])
 
 /**
  * Unfolds a record into a list of key/value pairs